pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.4,<3.13"
requests = "^2.31.0"
watchdog = "^3.0.0"

numpy = ">=1.26.0"
pandas = "^2.1.1"
streamlit = "^1.27.1"
streamlit-option-menu = "^0.3.6"
psycopg2 = "^2.9.8"
google-cloud-firestore = "^2.12.0"
oauth2client = "^4.1.3"
google-auth = "^2.23.2"
firebase-admin = "^6.2.0"
fastapi = "^0.103.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
